Add hierarchical expand/collapse to nodes, node visibility, title texts with scaling, line width scaling, small text svg simplification
As part of the master project Google Maps for Models, nodes have an expansion state, that can be set to expand/collapses according to zoom level and diagram dimensions, nodes that are not visible in the viewport can be collapsed, for collapsed nodes the title of the region can be scaled for readability as well as macro state labels, if there is just one in a node, line widths can be scaled to appear with constant width in diagram and small texts can be simplified with rectangles. This functionality can be configured in the Render Options of the diagram.